+++ packages/turbo_bridge/lib/src/services/app_info_service.dart
@@ -33,7 +33,7 @@ class AppInfoService {
       'darkMode':
           binding.platformDispatcher.platformBrightness == ui.Brightness.dark,
       'currentRoute': currentRoute,
-      'bridgeVersion': '0.1.1',
+      'bridgeVersion': '0.1.2',
       'locale': binding.platformDispatcher.locale.toString(),
     };
   }